refactor: relocate utils/image.ts to core/utils, fix 3 any-type violations
- Moved src/utils/image.ts → src/core/utils/image.ts (8 imports updated,
  orphan utils/ directory removed)
- RoomGrid.vue: canUpgrade/getUpgradeCost (room: any) → (room: Room)
- TrainingRoomModal.vue: type Room/Dweller = any → proper imports
- DwellersView.vue: getRelevantStatForRoom (dweller: any, room: any)
  → (dweller: Dweller, room: Room)
