+++ frontend/src/core/utils/image.ts
