+++ frontend/src/modules/chat/composables/useChatMessages.ts
@@ -3,7 +3,7 @@ import { onKeyStroke } from '@vueuse/core'
 import apiClient from '@/core/plugins/axios'
 import type { useChatWebSocket } from '@/core/composables/useWebSocket'
 import { handleStoreError } from '@/core/utils/errorHandler'
-import { normalizeImageUrl } from '@/utils/image'
+import { normalizeImageUrl } from '@/core/utils/image'
 import type { ChatMessageDisplay } from '@/modules/chat/models/chat'
 
 export interface UseChatMessagesOptions {
